requirements.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn[standard]==0.32.1
pydantic==2.10.4

# Data processing
pandas==2.2.3
numpy==1.26.4
joblib==1.5.1

# Machine learning
scikit-learn==1.5.2
xgboost==2.1.1

# Timezone handling
pytz==2024.2

# Environment variables
python-dotenv==1.0.1

# LangChain and AI dependencies
langchain==0.3.27
langchain-openai==0.2.14
langchain-community==0.3.27
langchain-chroma==0.2.5
openai==1.58.1

# Vector database
chromadb==1.0.16

# Document processing
pymupdf==1.24.3
tiktoken==0.8.0

# HTTP client
httpx==0.28.1
aiohttp==3.12.15

# Retry logic
tenacity==8.2.3

# Utilities
python-multipart==0.0.20
bs4